[feature/feedback-form-display] - Fixed Display.js to work with Card component
